lr9/requirements.txt
psycopg2-binary==2.9.9
faker==24.8.0
tabulate==0.9.0
python-dotenv==1.0.0